split bbUtil into lib package of modules: discordUtil, emojis, jsonHandler, pathfinding, stringTyping and timeUtil.
