+++ BB/bbConfig/bbConfig.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 import math, random, pprint
-from ..bbUtil import dumbEmoji
+from ..lib.emojis import dumbEmoji
 
 ##### UTIL #####
