+++ BB/bbObjects/battles/DuelRequest.py
@@ -1,9 +1,11 @@
-from ... import bbUtil, bbGlobals
+from ... import lib, bbGlobals
 from ...bbConfig import bbConfig
 from discord import Embed, User, Message
 from .. import bbUser
 from ...scheduling import TimedTask
 from .. import bbGuild
+from ..items import bbShip
+import random
 
 
 def makeDuelStatsEmbed(duelResults : dict, targetUser : bbUser.bbUser, sourceUser : bbUser.bbUser) -> Embed:
@@ -61,6 +63,97 @@ class DuelRequest:
 
 
 # ⚠⚠⚠ THIS FUNCTION IS MARKED FOR CHANGE
+def fightShips(ship1 : bbShip.bbShip, ship2 : bbShip.bbShip, variancePercent : float) -> dict:
+    """Simulate a duel between two ships.
+    Returns a dictionary containing statistics about the duel, as well as a reference to the winning ship. 
+
+    :param bbShip ship1: One of the ships partaking in the duel 
+    :param bbShip ship2: One of the ships partaking in the duel
+    :param float variancePercent: The amount of random variance to apply to ship statistics, as a float percentage (e.g 0.5 for 50% random variance lll)
+    :return: A dictionary containing statistics about the duel, as well as a reference to the winning ship.
+    :rtype: dict
+    """
+
+    # Fetch ship total healths
+    ship1HP = ship1.getArmour() + ship1.getShield()
+    ship2HP = ship2.getArmour() + ship2.getShield()
+
+    # Vary healths by +=variancePercent
+    ship1HPVariance = ship1HP * variancePercent
+    ship2HPVariance = ship2HP * variancePercent
+    ship1HPVaried = random.randint(int(ship1HP - ship1HPVariance), int(ship1HP + ship1HPVariance))
+    ship2HPVaried = random.randint(int(ship2HP - ship2HPVariance), int(ship2HP + ship2HPVariance))
+
+    # Fetch ship total DPSs
+    ship1DPS = ship1.getDPS()
+    ship2DPS = ship2.getDPS()
+
+    if ship1DPS == 0:
+        if ship2DPS == 0:
+            return {"winningShip": None,
+            "ship1":{"health":{"stock":ship1HP, "varied":ship1HP},
+                    "DPS": {"stock":ship1DPS, "varied":ship1DPS},
+                    "TTK": -1},
+            "ship2":{"health":{"stock":ship2HP, "varied":ship2HP},
+                    "DPS": {"stock":ship2DPS, "varied":ship2DPS},
+                    "TTK": -1}}
+        return {"winningShip": ship2,
+            "ship1":{"health":{"stock":ship1HP, "varied":ship1HP},
+                    "DPS": {"stock":ship1DPS, "varied":ship1DPS},
+                    "TTK": round(ship1HP / ship2DPS, 2)},
+            "ship2":{"health":{"stock":ship2HP, "varied":ship2HP},
+                    "DPS": {"stock":ship2DPS, "varied":ship2DPS},
+                    "TTK": -1}}
+    if ship2DPS == 0:
+        if ship1DPS == 0:
+            return {"winningShip": None,
+            "ship1":{"health":{"stock":ship1HP, "varied":ship1HP},
+                    "DPS": {"stock":ship1DPS, "varied":ship1DPS},
+                    "TTK": -1},
+            "ship2":{"health":{"stock":ship2HP, "varied":ship2HP},
+                    "DPS": {"stock":ship2DPS, "varied":ship2DPS},
+                    "TTK": -1}}
+        return {"winningShip": ship1,
+            "ship1":{"health":{"stock":ship1HP, "varied":ship1HP},
+                    "DPS": {"stock":ship1DPS, "varied":ship1DPS},
+                    "TTK": -1},
+            "ship2":{"health":{"stock":ship2HP, "varied":ship2HP},
+                    "DPS": {"stock":ship2DPS, "varied":ship2DPS},
+                    "TTK": round(ship2HP / ship1DPS, 2)}}
+
+    # Vary DPSs by +=variancePercent
+    ship1DPSVariance = ship1DPS * variancePercent
+    ship2DPSVariance = ship2DPS * variancePercent
+    ship1DPSVaried = random.randint(int(ship1DPS - ship1DPSVariance), int(ship1DPS + ship1DPSVariance))
+    ship2DPSVaried = random.randint(int(ship2DPS - ship2DPSVariance), int(ship2DPS + ship2DPSVariance))
+
+    # Handling to be implemented
+    # ship1Handling = ship1.getHandling()
+    # ship2Handling = ship2.getHandling()
+    # ship1HandlingPenalty = 
+
+    # Calculate ship TTKs
+    ship1TTK = ship1HPVaried / ship2DPSVaried
+    ship2TTK = ship2HPVaried / ship1DPSVaried
+
+    # Return the ship with the longest TTK as the winner
+    if ship1TTK > ship2TTK:
+        winningShip = ship1
+    elif ship2TTK > ship1TTK:
+        winningShip = ship2
+    else:
+        winningShip = None
+    
+    return {"winningShip":winningShip,
+            "ship1":{"health":{"stock":ship1HP, "varied":ship1HPVaried},
+                    "DPS": {"stock":ship1DPS, "varied":ship1DPSVaried},
+                    "TTK": ship1TTK},
+            "ship2":{"health":{"stock":ship2HP, "varied":ship2HPVaried},
+                    "DPS": {"stock":ship2DPS, "varied":ship2DPSVaried},
+                    "TTK": ship2TTK}}
+
+
+# ⚠⚠⚠ THIS FUNCTION IS MARKED FOR CHANGE
 async def fightDuel(sourceUser : User, targetUser : User, duelReq : DuelRequest, acceptMsg : Message) -> dict:
     """Simulate a duel between two users.
     Returns a dictionary containing statistics about the duel, as well as references to the winning and losing bbUsers. 
@@ -80,7 +173,7 @@ async def fightDuel(sourceUser : User, targetUser : User, duelReq : DuelRequest,
 
     # fight = ShipFight.ShipFight(sourceBBUser.activeShip, targetBBUser.activeShip)
     # duelResults = fight.fightShips(bbConfig.duelVariancePercent)
-    duelResults = bbUtil.fightShips(
+    duelResults = fightShips(
         sourceBBUser.activeShip, targetBBUser.activeShip, bbConfig.duelVariancePercent)
     winningShip = duelResults["winningShip"]
 
@@ -102,7 +195,7 @@ async def fightDuel(sourceUser : User, targetUser : User, duelReq : DuelRequest,
     if winningBBUser is None:
         await acceptMsg.channel.send(":crossed_swords: **Stalemate!** " + str(targetUser) + " and " + sourceUser.mention + " drew in a duel!")
         if acceptMsg.guild.get_member(targetUser.id) is None:
-            targetDCGuild = bbUtil.findBBUserDCGuild(targetBBUser)
+            targetDCGuild = lib.discordUtil.findBBUserDCGuild(targetBBUser)
             if targetDCGuild is not None:
                 targetBBGuild = bbGlobals.guildsDB.getGuild(targetDCGuild.id)
                 if targetBBGuild.hasPlayChannel():
@@ -136,7 +229,7 @@ async def fightDuel(sourceUser : User, targetUser : User, duelReq : DuelRequest,
 
         if acceptMsg.guild.get_member(winningBBUser.id) is None:
             await acceptMsg.channel.send(":crossed_swords: **Fight!** " + str(bbGlobals.client.get_user(winningBBUser.id)) + " beat " + bbGlobals.client.get_user(losingBBUser.id).mention + " in a duel!\n" + creditsMsg, embed=statsEmbed)
-            winnerDCGuild = bbUtil.findBBUserDCGuild(winningBBUser)
+            winnerDCGuild = lib.discordUtil.findBBUserDCGuild(winningBBUser)
             if winnerDCGuild is not None:
                 winnerBBGuild = bbGlobals.guildsDB.getGuild(winnerDCGuild.id)
                 if winnerBBGuild.hasPlayChannel():
@@ -144,7 +237,7 @@ async def fightDuel(sourceUser : User, targetUser : User, duelReq : DuelRequest,
         else:
             if acceptMsg.guild.get_member(losingBBUser.id) is None:
                 await acceptMsg.channel.send(":crossed_swords: **Fight!** " + bbGlobals.client.get_user(winningBBUser.id).mention + " beat " + str(bbGlobals.client.get_user(losingBBUser.id)) + " in a duel!\n" + creditsMsg, embed=statsEmbed)
-                loserDCGuild = bbUtil.findBBUserDCGuild(losingBBUser)
+                loserDCGuild = lib.discordUtil.findBBUserDCGuild(losingBBUser)
                 if loserDCGuild is not None:
                     loserBBGuild = bbGlobals.guildsDB.getGuild(loserDCGuild.id)
                     if loserBBGuild.hasPlayChannel():
@@ -178,7 +271,7 @@ async def rejectDuel(duelReq : DuelRequest, rejectMsg : Message, challenger : Us
 
     await rejectMsg.channel.send(":white_check_mark: You have rejected **" + str(challenger) + "**'s duel challenge.")
     if rejectMsg.guild.get_member(duelReq.sourceBBUser.id) is None:
-        targetDCGuild = bbUtil.findBBUserDCGuild(duelReq.sourceBBUser.id)
+        targetDCGuild = lib.discordUtil.findBBUserDCGuild(duelReq.sourceBBUser.id)
         if targetDCGuild is not None:
             targetBBGuild = bbGlobals.guildsDB.getGuild(targetDCGuild.id)
             if targetBBGuild.hasPlayChannel():
