+++ BB/reactionMenus/ReactionDuelChallengeMenu.py
@@ -1,6 +1,6 @@
 from . import ReactionMenu
 from ..bbConfig import bbConfig
-from .. import bbGlobals, bbUtil
+from .. import bbGlobals, lib
 from discord import Colour, NotFound, HTTPException, Forbidden, Message, Embed
 from datetime import datetime
 from ..scheduling import TimedTask
@@ -103,7 +103,7 @@ async def fromDict(rmDict : dict) -> ReactionDuelChallengeMenu:
 
     reactionRoles = {}
     for reaction in rmDict["options"]:
-        reactionRoles[bbUtil.dumbEmojiFromStr(reaction)] = dcGuild.get_role(rmDict["options"][reaction]["role"])
+        reactionRoles[bblib.emojisUtil.dumbEmojiFromStr(reaction)] = dcGuild.get_role(rmDict["options"][reaction]["role"])
 
     timeoutTT = None
     if "timeout" in rmDict:
