+++ BB/reactionMenus/ReactionMenu.py
@@ -3,7 +3,7 @@
 import inspect
 from discord import Embed, Colour, NotFound, HTTPException, Forbidden, Member, Message
 from ..bbConfig import bbConfig
-from .. import bbGlobals, bbUtil
+from .. import bbGlobals, lib
 from abc import ABC, abstractmethod
 
 
@@ -62,7 +62,7 @@ class ReactionMenuOption:
     :var name: The name of this option, as displayed in the menu embed.
     :vartype name: str
     :var emoji: The emoji that a user must react with to trigger this option
-    :vartype emoji: bbUtil.dumbEmoji
+    :vartype emoji: lib.emojis.dumbEmoji
     :var addFunc: The function to call when this option is added by a user
     :vartype addFunc: function
     :var removeFunc: The function to call when this option is removed by a user
@@ -83,10 +83,10 @@ class ReactionMenuOption:
     :vartype removeHasArgs: bool
     """
 
-    def __init__(self, name : str, emoji : bbUtil.dumbEmoji, addFunc=None, addArgs=None, removeFunc=None, removeArgs=None):
+    def __init__(self, name : str, emoji : lib.emojis.dumbEmoji, addFunc=None, addArgs=None, removeFunc=None, removeArgs=None):
         """
         :param str name: The name of this option, as displayed in the menu embed.
-        :param bbUtil.dumbEmoji emoji: The emoji that a user must react with to trigger this option
+        :param lib.emojis.dumbEmoji emoji: The emoji that a user must react with to trigger this option
         :param function addFunc: The function to call when this option is added by a user
         :param function removeFunc: The function to call when this option is removed by a user
         :param addArgs: The arguments to pass to addFunc. No type checking is done on this parameter, but a dict is recommended as a close replacement for keyword args.
@@ -186,10 +186,10 @@ class NonSaveableReactionMenuOption(ReactionMenuOption):
     Instead, inherit directly from ReactionMenuOption or another suitable subclass that is not marked as unsaveable.
     """
 
-    def __init__(self, name : str, emoji : bbUtil.dumbEmoji, addFunc=None, addArgs=None, removeFunc=None, removeArgs=None):
+    def __init__(self, name : str, emoji : lib.emojis.dumbEmoji, addFunc=None, addArgs=None, removeFunc=None, removeArgs=None):
         """
         :param str name: The name of this option, as displayed in the menu embed.
-        :param bbUtil.dumbEmoji emoji: The emoji that a user must react with to trigger this option
+        :param lib.emojis.dumbEmoji emoji: The emoji that a user must react with to trigger this option
         :param function addFunc: The function to call when this option is added by a user
         :param function removeFunc: The function to call when this option is removed by a user
         :param addArgs: The arguments to pass to addFunc. No type checking is done on this parameter, but a dict is recommended as a close replacement for keyword args.
@@ -211,10 +211,10 @@ class DummyReactionMenuOption(ReactionMenuOption):
     """A reaction menu option with no function calls.
     A prime example is ReactionPollMenu, where adding and removing options need not have any functionality.
     """
-    def __init__(self, name : str, emoji : bbUtil.dumbEmoji):
+    def __init__(self, name : str, emoji : lib.emojis.dumbEmoji):
         """
         :param str name: The name of this option, as displayed in the menu embed.
-        :param bbUtil.dumbEmoji emoji: The emoji that a user must react with to trigger this option
+        :param lib.emojis.dumbEmoji emoji: The emoji that a user must react with to trigger this option
         """
         super(DummyReactionMenuOption, self).__init__(name, emoji)
 
@@ -258,7 +258,7 @@ class ReactionMenu:
     :var msg: the message where this menu is embedded
     :vartype msg: discord.Message
     :var options: A dictionary storing all of the menu's options and their behaviour
-    :vartype options: dict[bbUtil.dumbEmoji, ReactionMenuOption]
+    :vartype options: dict[bblib.emojisUtil.dumbEmoji, ReactionMenuOption]
     :var titleTxt: The content of the embed title
     :vartype titleTxt: str
     :var desc: he content of the embed description; appears at the top below the title
@@ -291,7 +291,7 @@ class ReactionMenu:
         """
         :param discord.Message msg: the message where this menu is embedded
         :param options: A dictionary storing all of the menu's options and their behaviour (Default {})
-        :type options: dict[bbUtil.dumbEmoji, ReactionMenuOption]
+        :type options: dict[lib.emojis.dumbEmoji, ReactionMenuOption]
         :param str titleTxt: The content of the embed title (Default "")
         :param str desc: he content of the embed description; appears at the top below the title (Default "")
         :param discord.Colour col: The colour of the embed's side strip (Default None)
@@ -306,11 +306,11 @@ class ReactionMenu:
         """
 
         if footerTxt == "" and timeout is not None:
-            footerTxt = "This menu will expire in " + bbUtil.td_format_noYM(timeout.expiryDelta) + "."
+            footerTxt = "This menu will expire in " + lib.discordUtil.td_format_noYM(timeout.expiryDelta) + "."
         
         # discord.message
         self.msg = msg
-        # Dict of bbUtil.dumbEmoji: ReactionMenuOption
+        # Dict of lib.emojis.dumbEmoji: ReactionMenuOption
         self.options = options
 
         self.titleTxt = titleTxt
@@ -326,17 +326,17 @@ class ReactionMenu:
         self.targetRole = targetRole
 
     
-    def hasEmojiRegistered(self, emoji : bbUtil.dumbEmoji) -> bool:
+    def hasEmojiRegistered(self, emoji : lib.emojis.dumbEmoji) -> bool:
         """Decide whether or not the given emoji is an option in this menu
 
-        :param bbUtil.dumbEmoji emoji: The emoji to test for membership
+        :param lib.emojis.dumbEmoji emoji: The emoji to test for membership
         :return: True if emoji is an option in this menu, False otherwise.
         :rtype: bool
         """
         return emoji in self.options
 
 
-    async def reactionAdded(self, emoji : bbUtil.dumbEmoji, member : Member):
+    async def reactionAdded(self, emoji : lib.emojis.dumbEmoji, member : Member):
         """Invoke an option's behaviour when it is selected by a user.
         This method should be called during your discord client's on_reaction_add or on_raw_reaction_add event.
         
@@ -346,7 +346,7 @@ class ReactionMenu:
         if member has targetRole.
         Both may be specified and required.
 
-        :param bbUtil.dumbEmoji emoji: The emoji that member reacted to the menu with
+        :param bbUtlib.emojisil.dumbEmoji emoji: The emoji that member reacted to the menu with
         :param discord.Member member: The member that added the emoji reaction
         :return: The result of the corresponding menu option's addFunc, if any
         """
@@ -360,7 +360,7 @@ class ReactionMenu:
         return await self.options[emoji].add(member)
 
     
-    async def reactionRemoved(self, emoji : bbUtil.dumbEmoji, member : Member):
+    async def reactionRemoved(self, emoji : lib.emojis.dumbEmoji, member : Member):
         """Invoke an option's behaviour when it is deselected by a user.
         This method should be called during your discord client's on_reaction_remove or on_raw_reaction_remove event.
         
@@ -370,7 +370,7 @@ class ReactionMenu:
         if member has targetRole.
         Both may be specified and required.
 
-        :param bbUtil.dumbEmoji emoji: The emoji reaction that member removed from the menu
+        :param lib.emojis.dumbEmoji emoji: The emoji reaction that member removed from the menu
         :param discord.Member member: The member that removed the emoji reaction
         :return: The result of the corresponding menu option's removeFunc, if any
         """
@@ -488,15 +488,15 @@ class CancellableReactionMenu(ReactionMenu):
     TODO: Currently, if a different cancelEmoji is specified, it will not be saved to file, and it must be specified again when reloading the menu. Just save the cancel emoji in the dict, outside of options
 
     :var cancelEmoji: The emoji used for the menu's cancel button.
-    :vartype cancelEmoji: bbUtil.dumbEmoji
+    :vartype cancelEmoji: lib.emojis.dumbEmoji
     """
     def __init__(self, msg : Message, options={}, cancelEmoji=bbConfig.defaultCancelEmoji,
                     titleTxt="", desc="", col=Embed.Empty, timeout=None, footerTxt="", img="", thumb="", icon="", authorName="", targetMember=None, targetRole=None):
         """
         :param discord.Message msg: the message where this menu is embedded
         :param options: A dictionary storing all of the menu's options and their behaviour (Default {})
-        :type options: dict[bbUtil.dumbEmoji, ReactionMenuOption]
-        :param bbUtil.dumbEmoji emoji: The emoji members should react with to cancel the menu. (Default bbConfig.defaultCancelEmoji)
+        :type options: dict[lib.emojis.dumbEmoji, ReactionMenuOption]
+        :param lib.emojis.dumbEmoji emoji: The emoji members should react with to cancel the menu. (Default bbConfig.defaultCancelEmoji)
         :param str titleTxt: The content of the embed title (Default "")
         :param str desc: he content of the embed description; appears at the top below the title (Default "")
         :param discord.Colour col: The colour of the embed's side strip (Default None)
