+++ BB/reactionMenus/ReactionPollMenu.py
@@ -1,6 +1,6 @@
 from . import ReactionMenu
 from ..bbConfig import bbConfig
-from .. import bbGlobals, bbUtil
+from .. import bbGlobals, lib
 from discord import Colour, NotFound, HTTPException, Forbidden, Emoji, PartialEmoji, Message, Embed
 from datetime import datetime
 from ..scheduling import TimedTask
@@ -31,9 +31,9 @@ async def printAndExpirePollResults(msgID : int):
 
     for reaction in menuMsg.reactions:
         if type(reaction.emoji) in [Emoji, PartialEmoji]:
-            currentEmoji = bbUtil.dumbEmoji(id=reaction.emoji.id)
+            currentEmoji = lib.emojis.dumbEmoji(id=reaction.emoji.id)
         else:
-            currentEmoji = bbUtil.dumbEmoji(unicode=reaction.emoji)
+            currentEmoji = lib.emojis.dumbEmoji(unicode=reaction.emoji)
 
         if currentEmoji is None:
             bbLogger.log("ReactPollMenu", "prtAndExpirePollResults", "Failed to fetch dumbEmoji for reaction: " + str(reaction), category="reactionMenus", eventType="INV_REACT")
@@ -110,7 +110,7 @@ class ReactionPollMenu(ReactionMenu.ReactionMenu):
         """
         :param discord.Message msg: the message where this menu is embedded
         :param options: A dictionary storing all of the poll options. Poll option behaviour functions are not called. TODO: Add reactionAdded/Removed overloads that just return and dont check anything
-        :type options: dict[bbUtil.dumbEmoji, ReactionMenuOption]
+        :type options: dict[lib.emojis.dumbEmoji, ReactionMenuOption]
         :param TimedTask timeout: The TimedTask responsible for expiring this menu
         :param discord.Member pollStarter: The user who started the poll, for printing in the menu embed. Optional. (Default None)
         :param bool multipleChoice: Whether to accept votes for multiple options from the same user, or to restrict users to one option vote per poll.
@@ -174,7 +174,7 @@ async def fromDict(rmDict : dict) -> ReactionPollMenu:
     """
     options = {}
     for emojiName in rmDict["options"]:
-        emoji = bbUtil.dumbEmojiFromStr(emojiName)
+        emoji = lib.emojis.dumbEmojiFromStr(emojiName)
         options[emoji] = ReactionMenu.DummyReactionMenuOption(rmDict["options"][emojiName], emoji)
 
     msg = await bbGlobals.client.get_guild(rmDict["guild"]).get_channel(rmDict["channel"]).fetch_message(rmDict["msg"])
