+++ BB/reactionMenus/ReactionRolePicker.py
@@ -1,6 +1,6 @@
 from . import ReactionMenu
 from ..bbConfig import bbConfig
-from .. import bbGlobals, bbUtil
+from .. import bbGlobals, lib
 from discord import Colour, NotFound, HTTPException, Forbidden, Guild, Role, Message
 from datetime import datetime
 from ..scheduling import TimedTask
@@ -63,7 +63,7 @@ class ReactionRolePickerOption(ReactionMenu.ReactionMenuOption):
     :vartype role: discord.Role 
     """
 
-    def __init__(self, emoji : bbUtil.dumbEmoji, role : Role, menu : ReactionMenu.ReactionMenu):
+    def __init__(self, emoji : lib.emojis.dumbEmoji, role : Role, menu : ReactionMenu.ReactionMenu):
         self.role = role
         super(ReactionRolePickerOption, self).__init__(self.role.name, emoji, addFunc=giveRole, addArgs=(menu.dcGuild, self.role, menu.msg.id), removeFunc=removeRole, removeArgs=(menu.dcGuild, self.role, menu.msg.id))
 
@@ -88,12 +88,12 @@ class ReactionRolePicker(ReactionMenu.ReactionMenu):
     TODO: replace dcGuild param with extracting msg.guild
     """
 
-    def __init__(self, msg : Message, reactionRoles : Dict[bbUtil.dumbEmoji, Role], dcGuild : Guild, titleTxt="**Role Menu**", desc="React for your desired role!", col=None, timeout=None, footerTxt="", img="", thumb="", icon="", authorName="", targetMember=None, targetRole=None):
+    def __init__(self, msg : Message, reactionRoles : Dict[lib.emojis.dumbEmoji, Role], dcGuild : Guild, titleTxt="**Role Menu**", desc="React for your desired role!", col=None, timeout=None, footerTxt="", img="", thumb="", icon="", authorName="", targetMember=None, targetRole=None):
         # TODO: Stop taking dcGuild, and instead extract dcGuild from msg.guild
         """
         :param discord.Message msg: the message where this menu is embedded
         :param reactionRoles: A dictionary where keys are emojis and values are the roles to grant/remove when adding/removing the emoji 
-        :type reactionRoles: dict[bbUtil.dumbEmoji, discord.Role]
+        :type reactionRoles: dict[lib.emojis.dumbEmoji, discord.Role]
         :param discord.Guild dcGuild: The guild where this menu is contained TODO: Remove and replace with extracting msg.guild
         :param str titleTxt: The content of the embed title (Default "**Role Menu**")
         :param str desc: he content of the embed description; appears at the top below the title (Default "React for your desired role!")
@@ -141,7 +141,7 @@ async def fromDict(rmDict : dict) -> ReactionRolePicker:
 
     reactionRoles = {}
     for reaction in rmDict["options"]:
-        reactionRoles[bbUtil.dumbEmojiFromStr(reaction)] = dcGuild.get_role(rmDict["options"][reaction]["role"])
+        reactionRoles[lib.emojis.dumbEmojiFromStr(reaction)] = dcGuild.get_role(rmDict["options"][reaction]["role"])
 
     timeoutTT = None
     if "timeout" in rmDict:
